+++ packages/core/src/components/bal-button/bal-button.tsx
@@ -1,7 +1,17 @@
-import { Component, h, Prop, Host, Event, EventEmitter, ComponentInterface, Element, State } from '@stencil/core'
+import {
+  Component,
+  h,
+  Prop,
+  Host,
+  Event,
+  EventEmitter,
+  ComponentInterface,
+  Element,
+  State,
+  Listen,
+} from '@stencil/core'
 import { Attributes, inheritAttributes } from '../../utils/attributes'
-import { rLCP } from '../../utils/helpers'
-import { ListenTo } from '../../utils/listen'
+import { rOnLoad } from '../../utils/helpers'
 
 @Component({
   tag: 'bal-button',
@@ -105,6 +115,11 @@ export class Button implements ComponentInterface {
   @Prop() rounded = false
 
   /**
+   * If `true` the button is a popup.
+   */
+  @Prop() balPopup = undefined
+
+  /**
    * Name of the left button icon
    */
   @Prop() icon = ''
@@ -159,7 +174,7 @@ export class Button implements ComponentInterface {
    */
   @Event() balDidRender!: EventEmitter<BalEvents.BalButtonDidRenderDetail>
 
-  @ListenTo('click', { capture: true, target: 'document' })
+  @Listen('click', { capture: true, target: 'document' })
   listenOnClick(ev: UIEvent) {
     if (this.disabled && ev.target && ev.target === this.el) {
       ev.preventDefault()
@@ -168,7 +183,12 @@ export class Button implements ComponentInterface {
   }
 
   componentDidLoad(): void {
-    rLCP(() => (this.isLargestContentPaintDone = true))
+    rOnLoad(() => (this.isLargestContentPaintDone = true))
+    if (this.el.getAttribute('bal-popup') && !this.aria?.haspopup) {
+      this.aria = {
+        haspopup: 'true',
+      }
+    }
   }
 
   componentWillLoad() {
@@ -178,9 +198,9 @@ export class Button implements ComponentInterface {
       'aria-controls',
       'aria-hidden',
       'tabindex',
+      'aria-haspopup',
     ])
   }
-
   componentDidRender() {
     this.balDidRender.emit()
   }
@@ -292,6 +312,7 @@ export class Button implements ComponentInterface {
       'aria-label':
         this.aria?.label || this.inheritAttributes['aria-label'] || this.aria?.title || this.inheritAttributes['title'],
       'aria-controls': this.aria?.controls || this.inheritAttributes['aria-controls'],
+      'aria-haspopup': this.aria?.haspopup || this.inheritAttributes['aria-haspopup'],
     }
 
     return (