+++ packages/core/src/components/bal-option-list/bal-option-list.tsx
@@ -1,4 +1,4 @@
-import { Component, h, ComponentInterface, Host, Element, Prop, Watch, Method, State } from '@stencil/core'
+import { Component, h, ComponentInterface, Host, Element, Prop, Watch, Method, State, Listen } from '@stencil/core'
 import isNil from 'lodash.isnil'
 import { Attributes, inheritAttributes } from '../../utils/attributes'
 import { BEM } from '../../utils/bem'
@@ -7,7 +7,6 @@ import { Loggable, Logger, LogInstance } from '../../utils/log'
 import { includes, startsWith } from '../bal-select/utils/utils'
 import { BalAriaForm, defaultBalAriaForm } from '../../utils/form'
 import { BalOption } from '../../utils/dropdown'
-import { ListenTo } from '../../utils/listen'
 
 @Component({
   tag: 'bal-option-list',
@@ -95,7 +94,7 @@ export class OptionList implements ComponentInterface, Loggable {
    * ------------------------------------------------------
    */
 
-  @ListenTo('balOptionFocus', { passive: true })
+  @Listen('balOptionFocus', { passive: true })
   listenToMouseEnter(ev: BalEvents.BalOptionFocus) {
     const options = this.options
     const indexToFocus = this.getOptionIndex(options, ev.detail.value)
@@ -104,7 +103,7 @@ export class OptionList implements ComponentInterface, Loggable {
     }
   }
 
-  @ListenTo('balOptionChange', { passive: false })
+  @Listen('balOptionChange', { passive: false })
   listenToOptionChange({ detail }: BalEvents.BalOptionFocus) {
     if (!this.multiple) {
       this.options.filter(option => option.value !== detail.value).forEach(option => (option.selected = false))