+++ packages/core/src/components/bal-tabs/components/tab-button.tsx
@@ -16,7 +16,7 @@ export interface TabButtonProps {
   isAccordionOpen: boolean
   isLinkList: boolean
   inverted: boolean
-  expanded: boolean
+  isExpanded: boolean
   spaceless: boolean
   clickable: boolean
   iconPosition: BalProps.BalTabsIconPosition
@@ -37,7 +37,7 @@ export const TabButton: FunctionalComponent<TabButtonProps> = ({
   isAccordionOpen,
   isLinkList,
   inverted,
-  expanded,
+  isExpanded,
   spaceless,
   clickable,
   iconPosition,
@@ -98,7 +98,7 @@ export const TabButton: FunctionalComponent<TabButtonProps> = ({
         ...bemEl.modifier('clickable').class(clickable),
         ...bemEl.modifier('accordion').class(accordion),
         ...bemEl.modifier('inverted').class(inverted),
-        ...bemEl.modifier('expanded').class(expanded),
+        ...bemEl.modifier('expanded').class(isExpanded),
         ...bemEl.modifier('spaceless').class(spaceless),
         ...bemEl.modifier('first').class(isFirst),
         ...bemEl.modifier('last').class(isLast),
@@ -136,6 +136,7 @@ export const TabButton: FunctionalComponent<TabButtonProps> = ({
         item={item}
         isMobile={isMobile}
         isVertical={isVertical}
+        isExpanded={isExpanded}
         hasBubble={hasLabelBubble}
         inverted={inverted}
         context={context}