+++ packages/core/src/utils/resize/resize.listener.ts
@@ -5,13 +5,16 @@ import { BalResizeInfo } from './resize.interfaces'
 
 export class BalResizeListener<TObserver> extends ListenerAbstract<TObserver, BalResizeInfo> {
   private resizeObserver: ResizeObserver | undefined
-  private debouncedNotify = debounce(() => this.notify(), 10)
+  private debouncedNotify = debounce((info: BalResizeInfo) => this.notify(info), 42)
+  private lastWidth: number | undefined
+  private lastHeight: number | undefined
 
   connect(el: HTMLElement): void {
     super.connect(el)
     if (typeof ResizeObserver === 'undefined') {
       return
     }
+
     if (this.resizeObserver !== undefined) {
       this.resizeObserver?.disconnect()
       this.resizeObserver = undefined
@@ -23,10 +26,28 @@ export class BalResizeListener<TObserver> extends ListenerAbstract<TObserver, Ba
           if (!Array.isArray(entries) || !entries.length) {
             return
           }
-          this.debouncedNotify()
+          const entry = entries[0]
+
+          if (this.lastWidth === undefined && this.lastHeight === undefined) {
+            this.lastWidth = entry.contentRect.width
+            this.lastHeight = entry.contentRect.height
+          } else {
+            const widthChanged = this.lastWidth !== entry.contentRect.width
+            const heightChanged = this.lastHeight !== entry.contentRect.height
+
+            if (widthChanged || heightChanged) {
+              this.debouncedNotify({
+                width: widthChanged,
+                height: heightChanged,
+              })
+              this.lastWidth = entry.contentRect.width
+              this.lastHeight = entry.contentRect.height
+            }
+          }
         })
       }
     })
+
     this.resizeObserver.observe(el)
   }