+++ packages/core/src/utils/swiper/controls/large-control.tsx
@@ -24,7 +24,7 @@ export const LargeControl: FunctionalComponent<LargeControlProps> = ({
   leftControlTitle,
   rightControlTitle,
 }) => {
-  const block = BEM.block('carousel')
+  const block = BEM.block('swiper')
   const controls = block.element('controls')
   const button = controls.element('button')
 
@@ -48,6 +48,7 @@ export const LargeControl: FunctionalComponent<LargeControlProps> = ({
         rounded
         inverted={inverted}
         onClick={() => onPreviousClick()}
+        tabindex={-1}
         disabled={isFirst}
         aria-hidden={isFirst && areControlsHidden ? 'true' : null}
         data-testid="bal-carousel-control-left"
@@ -65,6 +66,7 @@ export const LargeControl: FunctionalComponent<LargeControlProps> = ({
         rounded
         inverted={inverted}
         onClick={() => onNextClick()}
+        tabindex={-1}
         disabled={isLast}
         aria-hidden={isLast && areControlsHidden ? 'true' : null}
         data-testid="bal-carousel-control-right"